fixed a cast

+++ src/edu/austral/prog2_2018c2/Player.java
@@ -95,7 +95,7 @@ public class Player extends Sprite implements Commons {
     }
 
     public void giveSpecialPower() {
-        double randomNumber = Math.random()*101; //genera un numero aleatorio entre 0 y 100. El Math.random() solo genera numeros entre 0 y 1
+        int randomNumber = (int) Math.random()*101; //genera un numero aleatorio entre 0 y 100. El Math.random() solo genera numeros entre 0 y 1
 
         if(randomNumber < 10){
             freezePower();
